+++ src/Aks/Aks.Test/ScenarioTests/KubernetesTests.cs
@@ -205,5 +205,12 @@ namespace Commands.Aks.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-EnableOidcIssuer");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestOutboundType()
+        {
+            TestRunner.RunTestScript("Test-OutboundType");
+        }
     }
 }