+++ src/Aks/Aks.Test/ScenarioTests/KubernetesTests.ps1
@@ -1194,3 +1194,22 @@ function Test-EnableOidcIssuer {
         Remove-AzResourceGroup -Name $resourceGroupName -Force
     }
 }
+
+function Test-OutboundType {
+    # Setup
+    $resourceGroupName = Get-RandomResourceGroupName
+    $kubeClusterName = Get-RandomClusterName
+    $location = 'eastus'
+    $nodeVmSize = "Standard_D2_v2"
+
+    try {
+        New-AzResourceGroup -Name $resourceGroupName -Location $location
+        
+        New-AzAksCluster -ResourceGroupName $resourceGroupName -Name $kubeClusterName -NodeVmSize $nodeVmSize -NodeCount 1 -OutboundType managedNATGateway
+        $cluster = Get-AzAksCluster -ResourceGroupName $resourceGroupName -Name $kubeClusterName
+        Assert-AreEqual 'managedNATGateway' $cluster.NetworkProfile.OutboundType
+    }
+    finally {
+        Remove-AzResourceGroup -Name $resourceGroupName -Force
+    }
+}